+++ src/System.Memory/src/System.Memory.csproj
@@ -100,6 +100,13 @@
     <Compile Include="..\Common\src\System\MutableDecimal.cs" />
   </ItemGroup>
   <ItemGroup Condition="'$(IsPartialFacadeAssembly)' != 'true'">
+    <Compile Include="$(CommonPath)\CoreLib\System\System\Memory.cs" />
+    <Compile Include="$(CommonPath)\CoreLib\System\System\MemoryDebugView.cs" />
+    <Compile Include="$(CommonPath)\CoreLib\System\System\ReadOnlyMemory.cs" />
+    <Compile Include="$(CommonPath)\CoreLib\System\System\Buffers\IRetainable.cs" />
+    <Compile Include="$(CommonPath)\CoreLib\System\System\Buffers\MemoryHandle.cs" />
+    <Compile Include="$(CommonPath)\CoreLib\System\System\Buffers\OwnedMemory.cs" />
+    <Compile Include="$(CommonPath)\CoreLib\System\System\Runtime\InteropServices\MemoryMarshal.cs" />
     <Compile Include="System\ReadOnlySpan.cs" />
     <Compile Include="System\Span.cs" />
     <Compile Include="System\SpanDebugView.cs" />
@@ -107,13 +114,6 @@
     <Compile Include="System\SpanHelpers.cs" />
     <Compile Include="System\Pinnable.cs" />
     <Compile Include="System\SpanHelpers.Clear.cs" />
-    <Compile Include="System\Memory.cs" />
-    <Compile Include="System\MemoryDebugView.cs" />
-    <Compile Include="System\ReadOnlyMemory.cs" />
-    <Compile Include="System\Buffers\IRetainable.cs" />
-    <Compile Include="System\Buffers\MemoryHandle.cs" />
-    <Compile Include="System\Buffers\OwnedMemory.cs" />
-    <Compile Include="System\Runtime\InteropServices\MemoryMarshal.cs" />
   </ItemGroup>
   <ItemGroup>
     <!-- Common or Common-branched source files -->
